Fix unit tests (move mocks into utils folder)
